Refactored each test type into its own class
