+++ auth-service/src/main/java/com/petclinic/auth/Role/RoleService.java
@@ -4,8 +4,6 @@ import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 
-import java.util.Optional;
-
 public interface RoleService {
 
     Role createRole(Role role);
